WIP web: Extract SpacePolicy to its own component
